verificacion/requirements.txt
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
coverage==7.1.0
exceptiongroup==1.1.0
Faker==16.8.1
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
future==0.18.3
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-pubsub==2.15.0
google-cloud-tasks==2.13.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
oauthlib==3.2.2
packaging==23.0
pandas
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.0
tldextract==3.4.0
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
